and again

+++ src/my_thing.hpp
@@ -504,7 +504,7 @@ using Thing = struct Thing {
   //
   // Increases per tick and when it reaches 1, allows the thing to move
   //
-  f16 thing_dt;
+  float thing_dt;
   //
   // Angle of movement.
   //
